put DziFiles::ActiveRecordCallbacks in it's own file
